requirements/requirements.txt
fastapi>=0.95.0
uvicorn>=0.23.0
sqlalchemy[asyncio]==2.0.20
pydantic>=1.10.0,<2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=6.0.0
httpx>=0.23.0
pymysql==1.1.1
asyncmy>=0.2.0